feat(cellpose): warm the weights cache in --init, share the GPU check with segment
Moves the cellpose helpers out of track.py into biahub/utils/cellpose.py, which
now states the two problems it exists to solve, and adds the piece the previous
commit left as a caveat.

`warm_cellpose_weights` runs from `track --init`, which the workflow executes once
on the head node before the per-position fan-out, and builds the model so cellpose
resolves and downloads whatever checkpoint it defaults to. That leaves the shared
cache in the state `stage_cellpose_weights` needs — populated, and populated once
instead of by every node in the fan-out. It is deliberately confined to the
`init_only` branch: workers run the same code path, and importing cellpose there
would freeze the weights directory before staging could redirect it. Warming is an
optimisation, so anything unexpected logs and leaves workers to download as before.

`biahub segment` now resolves its device through the same `cellpose_device`. It had
the same silent fallback in a different shape — `torch.device("cuda" if
torch.cuda.is_available() else "cpu")`, where `is_available()` returns False on a
failed CUDA init — and every job it submits requests `gpu:1` on the gpu partition
with no config knob for CPU, so an unusable GPU there is a broken allocation, not a
reason to run 130x slower. Its docstring was also sitting below the cellpose
import, where it was a no-op expression rather than `__doc__`; moving the import
fixes that.

Not staging weights in segment: it loads `path_to_model`, usually a path to a
user-trained model, and cellpose reads an existing path directly without consulting
its models directory. Staging would copy ~2.4 GB of unused built-in checkpoints to
every node for no gain.

Verified on Bruno: `track --init` reports `Cellpose weights ready:
.../cpsam_v2` against a warm cache in 19s, and against an empty
CELLPOSE_LOCAL_MODELS_PATH downloads 1.15 GB in 23s and then reports it, so a cold
shared cache is populated once at init rather than by every worker. Re-ran the
position from the incident after the move (job 35493397, gpu-g-2): weights staged,
`cellpose device: cuda:0`, cells per frame mean=34.5 min=19 max=48, unchanged.

Full suite: 159 passed, 1 skipped.
