+++ tests/test_cli/test_track_cli.py
@@ -1,11 +1,8 @@
-import getpass
 import os
-import sys
 
 import numpy as np
 import pandas as pd
 import pytest
-import torch
 import yaml
 
 from click.testing import CliRunner
@@ -13,13 +10,7 @@ from iohub.ngff import open_ome_zarr
 
 from biahub.cli.main import cli
 from biahub.settings import TrackingSettings, ZSlicing
-from biahub.track import (
-    _cellpose_device,
-    _init_output_plate,
-    _stage_cellpose_weights,
-    resolve_z_slice,
-    track,
-)
+from biahub.track import _init_output_plate, resolve_z_slice, track
 
 
 @pytest.fixture(scope="function")
@@ -509,74 +500,3 @@ def test_input_images_path_override(tmp_path, example_tracking_plate, monkeypatc
 
     assert result.exit_code == 0, result.output
     assert output_path.exists()
-
-
-def test_cellpose_device_cpu_when_not_requested():
-    assert _cellpose_device(gpu=False) == torch.device("cpu")
-
-
-def test_cellpose_device_raises_instead_of_falling_back(monkeypatch):
-    """An unusable GPU must fail loudly, not silently segment on the CPU."""
-
-    def unavailable(*args, **kwargs):
-        raise RuntimeError("CUDA error: all CUDA-capable devices are busy or unavailable")
-
-    monkeypatch.setattr(torch, "zeros", unavailable)
-    monkeypatch.setenv("CUDA_VISIBLE_DEVICES", "3")
-
-    with pytest.raises(RuntimeError, match="Refusing to fall back to the CPU") as excinfo:
-        _cellpose_device(gpu=True)
-
-    # The message has to carry enough to debug the node it happened on.
-    assert "CUDA_VISIBLE_DEVICES='3'" in str(excinfo.value)
-    assert "busy or unavailable" in str(excinfo.value)
-
-
-@pytest.fixture()
-def cellpose_cache(tmp_path, monkeypatch):
-    """A shared weights directory and an empty node-local scratch to stage into."""
-    # cellpose.models freezes its model directory at import time, so staging is a
-    # no-op once it is imported -- and another test may have imported it already.
-    monkeypatch.delitem(sys.modules, "cellpose.models", raising=False)
-
-    shared = tmp_path / "shared"
-    shared.mkdir()
-    (shared / "cpsam_v2").write_bytes(b"cpsam_v2 weights")
-    monkeypatch.setenv("CELLPOSE_LOCAL_MODELS_PATH", str(shared))
-    monkeypatch.setenv("TMPDIR", str(tmp_path / "scratch"))
-    return shared
-
-
-def test_stage_cellpose_weights_copies_to_scratch(cellpose_cache):
-    dest = _stage_cellpose_weights()
-
-    assert dest is not None
-    assert (dest / "cpsam_v2").read_bytes() == b"cpsam_v2 weights"
-    # cellpose reads the staged copy, and nothing partial is left behind.
-    assert os.environ["CELLPOSE_LOCAL_MODELS_PATH"] == str(dest)
-    assert sorted(path.name for path in dest.iterdir()) == ["cpsam_v2"]
-
-    # Staging again is a no-op now that the staging dir is the source.
-    assert _stage_cellpose_weights() == dest
-
-
-def test_stage_cellpose_weights_replaces_truncated_copy(cellpose_cache, tmp_path):
-    """A partial copy left behind by a killed task is replaced, not trusted."""
-    scratch = tmp_path / "scratch" / f"cellpose-models-{getpass.getuser()}"
-    scratch.mkdir(parents=True)
-    (scratch / "cpsam_v2").write_bytes(b"truncated")
-
-    dest = _stage_cellpose_weights()
-
-    assert dest == scratch
-    assert (dest / "cpsam_v2").read_bytes() == b"cpsam_v2 weights"
-
-
-def test_stage_cellpose_weights_skips_when_no_shared_cache(tmp_path, monkeypatch):
-    monkeypatch.delitem(sys.modules, "cellpose.models", raising=False)
-    monkeypatch.setenv("CELLPOSE_LOCAL_MODELS_PATH", str(tmp_path / "never_downloaded"))
-    monkeypatch.setenv("TMPDIR", str(tmp_path / "scratch"))
-
-    assert _stage_cellpose_weights() is None
-    # Left alone, so cellpose downloads to the shared cache as it always has.
-    assert os.environ["CELLPOSE_LOCAL_MODELS_PATH"] == str(tmp_path / "never_downloaded")
